fix: move ClientIpDep to fastapi/shared/dependencies.py after rebase
Main moved fastapi/utils.py to fastapi/shared/utils.py and introduced
shared/dependencies.py as the home for dependency functions. The rebase
kept this branch's imports pointing at the old module path, which broke
imports at startup (collection error in test_unlink.py).

- Move get_client_ip / ClientIpDep into shared/dependencies.py, next to
  get_fullpath, matching the module's stated purpose.
- Repoint link.py, importapi.py, and merge_authors.py imports.
